fix ADC issue with correct reading

+++ MCAL/02-ADC/inc/ADC_Interface.h
@@ -5,7 +5,7 @@
 #define ADC_INT_H
 
 extern void ADC_Init (void); // ADC Initialization And Enable
-extern uint8 ADC_u8ReadChannel (uint8  u8Channel);// Read From The ADC Channel
+extern uint16 ADC_u8ReadChannel (uint8  u8Channel);// Read From The ADC Channel
 
 
 #endif
